feat: improve handling of resource units
