+++ src/shared/helpers/resources.js
@@ -5,6 +5,7 @@ const SI_PREFIXES = {
   m: 1e-3,
   '': 1,
   k: 1e3,
+  K: 1e3,
   M: 1e6,
   G: 1e9,
   T: 1e12,
@@ -13,20 +14,28 @@ const SI_PREFIXES = {
 };
 
 const SI_PREFIXES_BINARY = {
+  '': 1,
   Ki: 2 ** 10,
   Mi: 2 ** 20,
   Gi: 2 ** 30,
   Ti: 2 ** 40,
   Pi: 2 ** 50,
+  Ei: 2 ** 60,
+};
+
+const MEMORY_PREFIXES = {
+  ...SI_PREFIXES,
+  ...SI_PREFIXES_BINARY,
 };
 
 /*
 More precise round method.
-Want 1.005 to be rounded to 1.01 we need to add Number.EPSILON to fix the float inaccuracy
+Fixes floating point precision issues (e.g., 1.005 rounds to 1.01)
  */
-const preciseRound = (num, places) =>
-  Math.round((num + Number.EPSILON) * Math.pow(10, places)) /
-  Math.pow(10, places);
+const preciseRound = (num, places) => {
+  const multiplier = Math.pow(10, places);
+  return Math.round((num + Number.EPSILON) * multiplier) / multiplier;
+};
 
 export function formatResourceUnit(
   amount = 0,
@@ -66,4 +75,37 @@ export function formatResourceUnit(
   });
 
   return output;
+}
+
+export function bytesToHumanReadable(bytes, { fixed = 0, unit = '' } = {}) {
+  return formatResourceUnit(bytes, true, { withoutSpace: true, fixed, unit });
+}
+
+export function cpusToHumanReadable(cpus, { fixed = 0, unit = '' } = {}) {
+  return formatResourceUnit(cpus, false, { withoutSpace: true, fixed, unit });
+}
+
+export function getCpus(cpuString) {
+  if (!cpuString || cpuString === '0') {
+    return 0;
+  }
+
+  const suffix = String(cpuString).slice(-1);
+
+  const suffixPower = SI_PREFIXES[suffix];
+  if (!suffixPower) {
+    return parseFloat(cpuString);
+  }
+
+  const number = String(cpuString).replace(suffix, '');
+  return number * suffixPower;
+}
+
+export function getBytes(memoryStr) {
+  if (!memoryStr) return 0;
+
+  const unit = String(memoryStr).match(/[a-zA-Z]+/g)?.[0];
+  const value = parseFloat(memoryStr);
+  const bytes = value * (MEMORY_PREFIXES[unit] || 1);
+  return bytes;
 }
